feat: restructuring folder
