+++ database/seeders/seeds/user_seed.go
@@ -1,0 +1,52 @@
+package seeds
+
+import (
+	"encoding/json"
+	"errors"
+	"io"
+	"os"
+
+	"github.com/Caknoooo/go-gin-clean-starter/database/entities"
+	"gorm.io/gorm"
+)
+
+func ListUserSeeder(db *gorm.DB) error {
+	jsonFile, err := os.Open("./migrations/json/users.json")
+	if err != nil {
+		return err
+	}
+
+	jsonData, err := io.ReadAll(jsonFile)
+	if err != nil {
+		return err
+	}
+
+	var listUser []entities.User
+	if err := json.Unmarshal(jsonData, &listUser); err != nil {
+		return err
+	}
+
+	hasTable := db.Migrator().HasTable(&entities.User{})
+	if !hasTable {
+		if err := db.Migrator().CreateTable(&entities.User{}); err != nil {
+			return err
+		}
+	}
+
+	for _, data := range listUser {
+		var user entities.User
+		err := db.Where(&entities.User{Email: data.Email}).First(&user).Error
+		if err != nil && !errors.Is(err, gorm.ErrRecordNotFound) {
+			return err
+		}
+
+		isData := db.Find(&user, "email = ?", data.Email).RowsAffected
+		if isData == 0 {
+			if err := db.Create(&data).Error; err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}
