+++ pkg/constants/common.go
@@ -1,0 +1,15 @@
+package constants
+
+const (
+	ENUM_ROLE_ADMIN = "admin"
+	ENUM_ROLE_USER  = "user"
+
+	ENUM_RUN_PRODUCTION = "production"
+	ENUM_RUN_TESTING    = "testing"
+
+	ENUM_PAGINATION_PER_PAGE = 10
+	ENUM_PAGINATION_PAGE     = 1
+
+	DB         = "db"
+	JWTService = "JWTService"
+)
