--- provider/user_provider.go
@@ -1,25 +1,0 @@
-package provider
-
-import (
-	"github.com/Caknoooo/go-gin-clean-starter/controller"
-	"github.com/Caknoooo/go-gin-clean-starter/repository"
-	"github.com/Caknoooo/go-gin-clean-starter/service"
-	"github.com/samber/do"
-	"gorm.io/gorm"
-)
-
-func ProvideUserDependencies(injector *do.Injector, db *gorm.DB, jwtService service.JWTService) {
-	// Repository
-	userRepository := repository.NewUserRepository(db)
-	refreshTokenRepository := repository.NewRefreshTokenRepository(db)
-
-	// Service
-	userService := service.NewUserService(userRepository, refreshTokenRepository, jwtService, db)
-
-	// Controller
-	do.Provide(
-		injector, func(i *do.Injector) (controller.UserController, error) {
-			return controller.NewUserController(userService), nil
-		},
-	)
-}
