+++ providers/core.go
@@ -1,0 +1,30 @@
+package providers
+
+import (
+	"github.com/Caknoooo/go-gin-clean-starter/config"
+	"github.com/Caknoooo/go-gin-clean-starter/modules/auth/service"
+	"github.com/Caknoooo/go-gin-clean-starter/pkg/constants"
+	"github.com/samber/do"
+	"gorm.io/gorm"
+)
+
+func InitDatabase(injector *do.Injector) {
+	do.ProvideNamed(injector, constants.DB, func(i *do.Injector) (*gorm.DB, error) {
+		return config.SetUpDatabaseConnection(), nil
+	})
+}
+
+func RegisterDependencies(injector *do.Injector) {
+	InitDatabase(injector)
+
+	do.ProvideNamed(injector, constants.JWTService, func(i *do.Injector) (service.JWTService, error) {
+		return service.NewJWTService(), nil
+	})
+
+	// Initialize
+	db := do.MustInvokeNamed[*gorm.DB](injector, constants.DB)
+	jwtService := do.MustInvokeNamed[service.JWTService](injector, constants.JWTService)
+
+	// Provide Dependencies
+	ProvideUserDependencies(injector, db, jwtService)
+}
