--- repository/user_repository.go
@@ -1,149 +1,0 @@
-package repository
-
-import (
-	"context"
-	"github.com/Caknoooo/go-gin-clean-starter/dto"
-	"github.com/Caknoooo/go-gin-clean-starter/entity"
-	"gorm.io/gorm"
-)
-
-type (
-	UserRepository interface {
-		Register(ctx context.Context, tx *gorm.DB, user entity.User) (entity.User, error)
-		GetAllUserWithPagination(
-			ctx context.Context,
-			tx *gorm.DB,
-			req dto.PaginationRequest,
-		) (dto.GetAllUserRepositoryResponse, error)
-		GetUserById(ctx context.Context, tx *gorm.DB, userId string) (entity.User, error)
-		GetUserByEmail(ctx context.Context, tx *gorm.DB, email string) (entity.User, error)
-		CheckEmail(ctx context.Context, tx *gorm.DB, email string) (entity.User, bool, error)
-		Update(ctx context.Context, tx *gorm.DB, user entity.User) (entity.User, error)
-		Delete(ctx context.Context, tx *gorm.DB, userId string) error
-	}
-
-	userRepository struct {
-		db *gorm.DB
-	}
-)
-
-func NewUserRepository(db *gorm.DB) UserRepository {
-	return &userRepository{
-		db: db,
-	}
-}
-
-func (r *userRepository) Register(ctx context.Context, tx *gorm.DB, user entity.User) (entity.User, error) {
-	if tx == nil {
-		tx = r.db
-	}
-
-	if err := tx.WithContext(ctx).Create(&user).Error; err != nil {
-		return entity.User{}, err
-	}
-
-	return user, nil
-}
-
-func (r *userRepository) GetAllUserWithPagination(
-	ctx context.Context,
-	tx *gorm.DB,
-	req dto.PaginationRequest,
-) (dto.GetAllUserRepositoryResponse, error) {
-	if tx == nil {
-		tx = r.db
-	}
-
-	var users []entity.User
-	var err error
-	var count int64
-
-	req.Default()
-
-	query := tx.WithContext(ctx).Model(&entity.User{})
-	if req.Search != "" {
-		query = query.Where("name LIKE ?", "%"+req.Search+"%")
-	}
-
-	if err := query.Count(&count).Error; err != nil {
-		return dto.GetAllUserRepositoryResponse{}, err
-	}
-
-	if err := query.Scopes(Paginate(req)).Find(&users).Error; err != nil {
-		return dto.GetAllUserRepositoryResponse{}, err
-	}
-
-	totalPage := TotalPage(count, int64(req.PerPage))
-	return dto.GetAllUserRepositoryResponse{
-		Users: users,
-		PaginationResponse: dto.PaginationResponse{
-			Page:    req.Page,
-			PerPage: req.PerPage,
-			Count:   count,
-			MaxPage: totalPage,
-		},
-	}, err
-}
-
-func (r *userRepository) GetUserById(ctx context.Context, tx *gorm.DB, userId string) (entity.User, error) {
-	if tx == nil {
-		tx = r.db
-	}
-
-	var user entity.User
-	if err := tx.WithContext(ctx).Where("id = ?", userId).Take(&user).Error; err != nil {
-		return entity.User{}, err
-	}
-
-	return user, nil
-}
-
-func (r *userRepository) GetUserByEmail(ctx context.Context, tx *gorm.DB, email string) (entity.User, error) {
-	if tx == nil {
-		tx = r.db
-	}
-
-	var user entity.User
-	if err := tx.WithContext(ctx).Where("email = ?", email).Take(&user).Error; err != nil {
-		return entity.User{}, err
-	}
-
-	return user, nil
-}
-
-func (r *userRepository) CheckEmail(ctx context.Context, tx *gorm.DB, email string) (entity.User, bool, error) {
-	if tx == nil {
-		tx = r.db
-	}
-
-	var user entity.User
-	if err := tx.WithContext(ctx).Where("email = ?", email).Take(&user).Error; err != nil {
-		return entity.User{}, false, err
-	}
-
-	return user, true, nil
-}
-
-func (r *userRepository) Update(ctx context.Context, tx *gorm.DB, user entity.User) (entity.User, error) {
-	if tx == nil {
-		tx = r.db
-	}
-
-	if err := tx.WithContext(ctx).Updates(&user).Error; err != nil {
-		return entity.User{}, err
-	}
-
-	return user, nil
-}
-
-func (r *userRepository) Delete(ctx context.Context, tx *gorm.DB, userId string) error {
-	if tx == nil {
-		tx = r.db
-	}
-
-	if err := tx.WithContext(ctx).Delete(&entity.User{}, "id = ?", userId).Error; err != nil {
-		return err
-	}
-
-	return nil
-}
