--- routes/router.go
@@ -1,10 +1,0 @@
-package routes
-
-import (
-	"github.com/gin-gonic/gin"
-	"github.com/samber/do"
-)
-
-func RegisterRoutes(server *gin.Engine, injector *do.Injector) {
-	User(server, injector)
-}
